Replaced test-utils crate with base_impl module
