+++ curves/bn254/Cargo.toml
@@ -11,16 +11,16 @@ license.workspace = true
 edition.workspace = true
 
 [dependencies]
+ark-bn254.workspace = true
 ark-ec.workspace = true
 ark-ff.workspace = true
-ark-std.workspace = true
-ark-bn254.workspace = true
 ark-models-ext.workspace = true
+ark-std.workspace = true
+ark-scale.workspace = true
 
 [dev-dependencies]
 ark-serialize.workspace = true
 ark-algebra-test-templates.workspace = true
-test-utils.workspace = true
 
 [features]
 default = [ "std" ]
@@ -29,7 +29,11 @@ std = [
   "ark-ec/std",
   "ark-ff/std",
   "ark-models-ext/std",
+  "ark-scale/std",
   "ark-serialize/std",
   "ark-std/std",
 ]
 parallel = [ "ark-ec/parallel", "ark-ff/parallel", "ark-std/parallel" ]
+# SCALE encoding related features for most of Arkworks types.
+# Defaults to "compressed" and "validated"
+scale-no-compress = []
