+++ curves/bn254/src/base_impl/mod.rs
@@ -34,6 +34,9 @@ use ark_scale::{
 };
 use ark_std::vec::Vec;
 
+use crate::CurveHooks;
+use ark_bn254::{g1::Config as ArkG1Config, g2::Config as ArkG2Config, Bn254 as ArkBn254};
+
 #[cfg(feature = "scale-no-compress")]
 const SCALE_COMPRESS: Compress = Compress::No;
 #[cfg(not(feature = "scale-no-compress"))]
@@ -94,6 +97,7 @@ pub fn msm_sw_generic<ExtCurve: SWCurveConfig, ArkCurve: SWCurveConfig>(
     res.try_transmute()
 }
 
+#[allow(dead_code)]
 pub fn msm_te_generic<ExtConfig: TECurveConfig, ArkConfig: TECurveConfig>(
     bases: &[TEAffine<ExtConfig>],
     scalars: &[ExtConfig::ScalarField],
@@ -116,6 +120,7 @@ pub fn mul_projective_sw_generic<ExtConfig: SWCurveConfig, ArkConfig: SWCurveCon
     res.try_transmute()
 }
 
+#[allow(dead_code)]
 pub fn mul_projective_te_generic<ExtConfig: TECurveConfig, ArkConfig: TECurveConfig>(
     base: &TEProjective<ExtConfig>,
     scalar: &[u64],
@@ -124,4 +129,51 @@ pub fn mul_projective_te_generic<ExtConfig: TECurveConfig, ArkConfig: TECurveCon
 
     let res = <ArkConfig as TECurveConfig>::mul_projective(&base, scalar);
     res.try_transmute()
+}
+
+pub struct TestHooks;
+
+pub type Bn254 = crate::Bn254<TestHooks>;
+pub type G1Projective = crate::G1Projective<TestHooks>;
+pub type G2Projective = crate::G2Projective<TestHooks>;
+pub type G1Affine = crate::G1Affine<TestHooks>;
+pub type G2Affine = crate::G2Affine<TestHooks>;
+pub type G1Config = crate::g1::Config<TestHooks>;
+pub type G2Config = crate::g2::Config<TestHooks>;
+
+impl CurveHooks for TestHooks {
+    fn bn254_multi_miller_loop(
+        g1: impl Iterator<Item = <Bn254 as Pairing>::G1Prepared>,
+        g2: impl Iterator<Item = <Bn254 as Pairing>::G2Prepared>,
+    ) -> Result<<Bn254 as Pairing>::TargetField, ()> {
+        multi_miller_loop_generic::<Bn254, ArkBn254>(g1, g2)
+    }
+
+    fn bn254_final_exponentiation(
+        target: <Bn254 as Pairing>::TargetField,
+    ) -> Result<<Bn254 as Pairing>::TargetField, ()> {
+        final_exponentiation_generic::<Bn254, ArkBn254>(target)
+    }
+
+    fn bn254_msm_g1(
+        bases: &[G1Affine],
+        scalars: &[<G1Config as CurveConfig>::ScalarField],
+    ) -> Result<G1Projective, ()> {
+        msm_sw_generic::<G1Config, ArkG1Config>(bases, scalars)
+    }
+
+    fn bn254_msm_g2(
+        bases: &[G2Affine],
+        scalars: &[<G2Config as CurveConfig>::ScalarField],
+    ) -> Result<G2Projective, ()> {
+        msm_sw_generic::<G2Config, ArkG2Config>(bases, scalars)
+    }
+
+    fn bn254_mul_projective_g1(base: &G1Projective, scalar: &[u64]) -> Result<G1Projective, ()> {
+        mul_projective_sw_generic::<G1Config, ArkG1Config>(base, scalar)
+    }
+
+    fn bn254_mul_projective_g2(base: &G2Projective, scalar: &[u64]) -> Result<G2Projective, ()> {
+        mul_projective_sw_generic::<G2Config, ArkG2Config>(base, scalar)
+    }
 }
