+++ curves/bn254/src/lib.rs
@@ -49,8 +49,10 @@
 )]
 #![allow(clippy::result_unit_err)]
 
+mod base_impl;
 mod curves;
 
 pub use ark_bn254::{fq, fq::*, fq12, fq12::*, fq2, fq2::*, fq6, fq6::*, fr, fr::*};
+pub use base_impl::TestHooks;
 
 pub use curves::*;
