--- test-utils/Cargo.toml
@@ -1,18 +1,0 @@
-[package]
-name = "test-utils"
-description = "Testing support crate"
-version.workspace = true
-edition.workspace = true
-publish = false
-
-[dependencies]
-ark-ec.workspace = true
-ark-std.workspace = true
-ark-scale.workspace = true
-
-[features]
-default = [ "std" ]
-std = [ "ark-ec/std", "ark-scale/std", "ark-std/std" ]
-# SCALE encoding related features for most of Arkworks types.
-# Defaults to "compressed" and "validated"
-scale-no-compress = []
